// src/app/types/Skjemanummer.ts
export enum Skjemanummer {
    ANNET = 'I000060',
    BEKREFTELSE_DELTAR_KVALIFISERINGSPROGRAM = 'I000051',
    BEKREFTELSE_FRA_STUDIESTED = 'I000061',
    DOKUMENTASJON_AV_TERMIN_ELLER_FØDSEL = 'I000041',
    DOK_AV_ALENEOMSORG = 'I000110',
    DOK_BEGRUNNELSE_SØKE_TILBAKE_I_TID = 'I000111',
    DOK_DELTAKELSE_I_INTRODUKSJONSPROGRAMMET = 'I000112',
    DOK_INNLEGGELSE = 'I000037',
    DOK_MILITÆR_SILVIL_TJENESTE = 'I000039',
    DOK_MORS_UTDANNING_ARBEID_SYKDOM = 'I000038',
    DOK_OVERFØRING_FOR_SYK = 'I000023',
    ETTERLØNN_ELLER_SLUTTVEDERLAG = 'I000044',
    INNTEKTSOPPLYSNINGER_FRILANS_ELLER_SELVSTENDIG = 'I000007',
    OMSORGSOVERTAKELSESDATO = 'I000042',
    TERMINBEKREFTELSE = 'I000062',
    HV_ØVELSE = 'I000116',
    NAV_TILTAK = 'I000117',
}